requirements.txt
google-cloud-texttospeech==2.17.1
python-dotenv==0.21.1
pydub==0.25.1

